add route and route stations crud and business logic

+++ database/migrations/2023_06_02_195010_create_route_stations_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('start_station_id')->constrained('stations')->cascadeOnDelete();
             $table->foreignId('end_station_id')->constrained('stations')->cascadeOnDelete();
             $table->integer('order');
-            $table->double('price',5,3)->default(0);
+            $table->float('price')->default(0);
             $table->timestamps();
         });
     }
